refactor(workflow): workflows are not just used for project installation anymore
