+++ app/src/plugins/aurelia-cli/index.ts
@@ -5,8 +5,8 @@ import {BasePlugin}             from '../base-plugin';
 import {Detection}              from './detection';
 import {Task}                   from '../task-manager/task';
 import {CommandRunner}          from '../task-manager/command-runner';
-import {Workflow}               from '../../project-installation/workflow';
-import {Step}                   from '../../project-installation/step';
+import {Workflow}               from '../workflow/workflow';
+import {Step}                   from '../workflow/step';
 import {PluginManager}          from '../../shared/plugin-manager';
 import {Project}                from '../../shared/project';
 import {CommandService}         from './command-service';
