+++ app/src/plugins/development/developer-modal.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../../project-installation/workflow-viewer"></require>
+  <require from="../workflow/workflow-viewer"></require>
 
   <ai-dialog class="developer-modal" width="600" title="Developer" height="500">
     <ai-dialog-body>
