+++ app/src/plugins/dotnet/index.ts
@@ -5,8 +5,8 @@ import {Detection}     from './detection';
 import {PluginManager} from '../../shared/plugin-manager';
 import {Project}       from '../../shared/project';
 import {Notification}  from '../../shared/notification';
-import {Workflow}      from '../../project-installation/workflow';
-import {Step}          from '../../project-installation/step';
+import {Workflow}      from '../workflow/workflow';
+import {Step}          from '../workflow/step';
 import {Errors}        from '../errors/errors';
 import {Task}          from '../task-manager/task';
 import {CommandRunner} from '../task-manager/command-runner';
