+++ app/src/plugins/gulp/index.ts
@@ -4,8 +4,8 @@ import {BasePlugin}    from '../base-plugin';
 import {Detection}     from './detection';
 import {Project}       from '../../shared/project';
 import {Task}          from '../task-manager/task';
-import {Workflow}      from '../../project-installation/workflow';
-import {Step}          from '../../project-installation/step';
+import {Workflow}      from '../workflow/workflow';
+import {Step}          from '../workflow/step';
 import {CommandRunner} from '../task-manager/command-runner';
 import {CommandService} from './command-service';
 import {CommandRunnerService} from '../task-manager/command-runner-service';
