+++ app/src/plugins/index.ts
@@ -1,5 +1,5 @@
 export function configure(aurelia) {
-  aurelia.feature('plugins/workflows');
+  aurelia.feature('plugins/workflow');
   aurelia.feature('plugins/app-launcher');
   aurelia.feature('plugins/npm');
   aurelia.feature('plugins/jspm');
