+++ app/src/plugins/jspm/index.ts
@@ -1,8 +1,8 @@
 import {autoinject}    from 'aurelia-framework';
 import {PluginManager} from '../../shared/plugin-manager';
 import {Project}       from '../../shared/project';
-import {Workflow}      from '../../project-installation/workflow';
-import {Step}          from '../../project-installation/step';
+import {Workflow}      from '../workflow/workflow';
+import {Step}          from '../workflow/step';
 import {Task}          from '../task-manager/task';
 import {BasePlugin}    from '../base-plugin';
 import {JSPMDetection} from './jspm-detection';
