+++ app/src/plugins/npm/index.ts
@@ -2,8 +2,8 @@ import {autoinject}       from 'aurelia-framework';
 import {Project}          from '../../shared/project';
 import {PluginManager}    from '../../shared/plugin-manager';
 import {ApplicationState} from '../../shared/application-state';
-import {Workflow}         from '../../project-installation/workflow';
-import {Step}             from '../../project-installation/step';
+import {Workflow}         from '../workflow/workflow';
+import {Step}             from '../workflow/step';
 import {Task}             from '../task-manager/task';
 import {NPMDetection}     from './npm-detection';
 import {BasePlugin}       from '../base-plugin';
