+++ app/src/scaffolding/post-create.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../project-installation/workflow-viewer"></require>
+  <require from="../plugins/workflow/workflow-viewer"></require>
   
   <p>Would you like to:</p>
   <workflow-viewer project.bind="project" checked-count.two-way="checkedCount" selected-tasks.two-way="selectedTasks" view-model.ref="workflowViewer"></workflow-viewer>
